server/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
gunicorn==21.2.0
langgraph==0.2.0
langchain==0.2.0
langchain-cohere>=0.2.0   
cohere>=5.1.5,<6.0.0
firebase-admin==6.3.0
python-dotenv==1.0.0
requests==2.31.0
python-dateutil==2.8.2
langchain-groq